Cleaned up FPGAThread, and refactored, putting FPGAThread, FPGAController, and FPGAFlags into package fpga. Deleted individual flag classes from fpga package.
